TYP: untyped decorators
