+++ pandas/core/base.py
@@ -198,7 +198,7 @@ class SelectionMixin:
         else:
             return self.obj[self._selection]
 
-    @cache_readonly
+    @property
     def ndim(self) -> int:
         return self._selected_obj.ndim
 
@@ -1314,53 +1314,6 @@ class IndexOpsMixin(OpsMixin):
         if dropna and isna(uniqs).any():
             n -= 1
         return n
-
-    @property
-    def is_unique(self) -> bool:
-        """
-        Return boolean if values in the object are unique.
-
-        Returns
-        -------
-        bool
-        """
-        return self.nunique(dropna=False) == len(self)
-
-    @property
-    def is_monotonic(self) -> bool:
-        """
-        Return boolean if values in the object are
-        monotonic_increasing.
-
-        Returns
-        -------
-        bool
-        """
-        from pandas import Index
-
-        return Index(self).is_monotonic
-
-    @property
-    def is_monotonic_increasing(self) -> bool:
-        """
-        Alias for is_monotonic.
-        """
-        # mypy complains if we alias directly
-        return self.is_monotonic
-
-    @property
-    def is_monotonic_decreasing(self) -> bool:
-        """
-        Return boolean if values in the object are
-        monotonic_decreasing.
-
-        Returns
-        -------
-        bool
-        """
-        from pandas import Index
-
-        return Index(self).is_monotonic_decreasing
 
     def memory_usage(self, deep=False):
         """
