+++ pandas/core/indexes/category.py
@@ -333,7 +333,7 @@ class CategoricalIndex(ExtensionIndex, accessor.PandasDelegate):
 
     # --------------------------------------------------------------------
 
-    @property
+    @cache_readonly
     def inferred_type(self) -> str:
         return "categorical"
