+++ pandas/core/indexes/datetimes.py
@@ -799,7 +799,7 @@ class DatetimeIndex(DatetimeTimedeltaMixin):
     def is_type_compatible(self, typ) -> bool:
         return typ == self.inferred_type or typ == "datetime"
 
-    @property
+    @cache_readonly
     def inferred_type(self) -> str:
         # b/c datetime is represented as microseconds since the epoch, make
         # sure we can't have ambiguous indexing
